requirements.txt
anthropic>=0.40.0       # includes AnthropicBedrock client
boto3>=1.34.0           # AWS SDK — needed for Bedrock credential chain
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
pypdf>=4.0.0
